Added more functionality to TradePlacementProcessor and performed refactoring
Added code to allow the TradePlacementProcessor class to determine the
results of trades according to a user-specified interval.
Renamed beans package to records package.
Moved the CamelExchangeFactory class to the utility package in the test
folder.
Fixed NullPointerException caused by missing headers in the
TradePlacementProcessor unit tests.
Did some light refactoring and corrected a spelling error.